requirements.txt
cat > requirements.txt << 'EOF'
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.54.4
python-dotenv==1.0.1
pydantic==2.9.2
EOF